+++ src/demos/DemoLinearColorspace.cxx
@@ -81,7 +81,7 @@ demo::DemoLinearColorspace::~DemoLinearColorspace()
 void demo::DemoLinearColorspace::OnWindowSizeChanged(int width, int height)
 {
     getRenderer().setViewport(0, 0, width, height);
-    m_camera.setAspect(width / static_cast<float>(height));
+    m_camera.setAspect(static_cast<float>(width) / static_cast<float>(height));
 }
 
 bool demo::DemoLinearColorspace::OnKeyPressed(int key, int scancode, int action, int mods)
