+++ src/demos/DemoPhongReflectionModel.cxx
@@ -63,7 +63,7 @@ demo::DemoPhongReflectionModel::~DemoPhongReflectionModel()
 void demo::DemoPhongReflectionModel::OnWindowSizeChanged(int width, int height)
 {
     getRenderer().setViewport(0, 0, width, height);
-    m_camera.setAspect(width / static_cast<float>(height));
+    m_camera.setAspect(static_cast<float>(width) / static_cast<float>(height));
 }
 
 bool demo::DemoPhongReflectionModel::OnKeyPressed(int key, int scancode, int action, int mods)
