+++ src/demos/DemoPhongReflectionModelTextured.cxx
@@ -73,7 +73,7 @@ demo::DemoPhongReflectionModelTextured::~DemoPhongReflectionModelTextured()
 void demo::DemoPhongReflectionModelTextured::OnWindowSizeChanged(int width, int height)
 {
     getRenderer().setViewport(0, 0, width, height);
-    m_camera.setAspect(width / static_cast<float>(height));
+    m_camera.setAspect(static_cast<float>(width) / static_cast<float>(height));
 }
 
 bool demo::DemoPhongReflectionModelTextured::OnKeyPressed(int key, int scancode, int action, int mods)
